2020.4.9 15 : 50 1.修改WarehouseController类，修改部分逻辑，并添加订单入库审核方法 2.修改CarDao，新增车辆到达修改信息方法 3.修改ManagementDao，新增完成订单以及获取承运车辆其他计划的方法 4.修改OrderDao，新增完成订单方法 5.修改WarehouseDao，新增按照名字获取仓库数据方法，以及入库后减少库存方法 6.修改ManagementObject，新增一个complete属性，表示计划完成 7.新增ToWarehouseObject 8.修改WarehouseUnit，修改总容量与剩余容量为浮点数 9.对应修改CarService和车辆映射并 10.对应修改ManagementService和映射表 11.对应修改OrderService和映射表 12.对应修改WarehouseService和映射表 13.新增ToWarehouseUtil工具类 14.修改WarehouseUtil工具类，添加削减库位库存方法 15.新增towarehouse.jsp 16.为新增的类和方法添加注释

+++ src/main/java/com/bishe/logistics_management/database/dataObject/WarehouseUnit.java
@@ -11,9 +11,9 @@ public class WarehouseUnit {
 
     private String belongPos;
 
-    private int capacity;//总容量
+    private float capacity;//总容量
 
-    private int size;//剩余容量
+    private float size;//剩余容量
 
     private String name;
 
@@ -42,19 +42,19 @@ public class WarehouseUnit {
         this.belongPos = belongPos;
     }
 
-    public int getCapacity() {
+    public float getCapacity() {
         return capacity;
     }
 
-    public void setCapacity(int capacity) {
+    public void setCapacity(float capacity) {
         this.capacity = capacity;
     }
 
-    public int getSize() {
+    public float getSize() {
         return size;
     }
 
-    public void setSize(int size) {
+    public void setSize(float size) {
         this.size = size;
     }
